Homework with network

+++ app/src/main/java/com/example/myownapplication/data/network/movie/MovieDto.kt
@@ -11,20 +11,20 @@ data class MovieTopRatedDto(
 
 @Serializable
 data class MovieDto(
-    val id: Int,
-    val title: String,
-    @SerialName("poster_path")
-    val posterPicture: String,
-    @SerialName("backdrop_path")
-    val backdropPicture: String,
+        val id: Long,
+        val title: String,
+        @SerialName("poster_path")
+        val posterPicture: String,
+        @SerialName("backdrop_path")
+        val backdropPicture: String,
 //            val runtime: Int,
-    @SerialName("genre_ids")
-    val genreIds: List<Int>,
+        @SerialName("genre_ids")
+        val genreIds: List<Int>,
 //            val actors: List<Int>,
-    @SerialName("vote_average")
-    val ratings: Float,
-    @SerialName("vote_count")
-    val votesCount: Int,
-    val overview: String,
-    val adult: Boolean
+        @SerialName("vote_average")
+        val ratings: Float,
+        @SerialName("vote_count")
+        val votesCount: Long,
+        val overview: String,
+        val adult: Boolean
 )
